Convert positions into a uint32 array

+++ src/mocks/MockCellar.sol
@@ -12,7 +12,7 @@ contract MockCellar is Cellar, Test {
     constructor(
         Registry _registry,
         ERC20 _asset,
-        uint256[] memory _positions,
+        uint32[] memory _positions,
         bytes[] memory _configurationData,
         string memory _name,
         string memory _symbol,
@@ -32,7 +32,7 @@ contract MockCellar is Cellar, Test {
     {}
 
     function depositIntoPosition(
-        uint256 position,
+        uint32 position,
         uint256 amount,
         address mintSharesTo
     ) external returns (uint256 shares) {
@@ -41,14 +41,14 @@ contract MockCellar is Cellar, Test {
         _mint(mintSharesTo, shares);
     }
 
-    function depositIntoPosition(uint256 position, uint256 amount) external returns (uint256 shares) {
+    function depositIntoPosition(uint32 position, uint256 amount) external returns (uint256 shares) {
         shares = _depositIntoPosition(position, amount);
 
         // Increase totalSupply by shares amount.
         stdstore.target(address(this)).sig(this.totalSupply.selector).checked_write(totalSupply() + shares);
     }
 
-    function _depositIntoPosition(uint256 position, uint256 amount) internal returns (uint256 shares) {
+    function _depositIntoPosition(uint32 position, uint256 amount) internal returns (uint256 shares) {
         ERC20 positionAsset = _assetOf(position);
 
         PriceRouter priceRouter = PriceRouter(registry.getAddress(2));
